+++ src/main/java/org/jallen/tyrael/entity/Work.java
@@ -1,0 +1,81 @@
+package org.jallen.tyrael.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+
+@Entity
+@Table(name = "work")
+public class Work {
+
+  @Id
+  @Column(name = "id", nullable = false)
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+
+  @Column(columnDefinition = "TEXT")
+  private String name;
+
+  @Column(columnDefinition = "TEXT")
+  private String sub;
+
+  @Column(columnDefinition = "TEXT")
+  private String technology;
+
+  @Column(columnDefinition = "TEXT")
+  private String company;
+
+  @Column(columnDefinition = "TEXT")
+  private String description;
+
+  public Long getId() {
+    return id;
+  }
+
+  public void setId(Long id) {
+    this.id = id;
+  }
+
+  public String getName() {
+    return name;
+  }
+
+  public void setName(String name) {
+    this.name = name;
+  }
+
+  public String getSub() {
+    return sub;
+  }
+
+  public void setSub(String sub) {
+    this.sub = sub;
+  }
+
+  public String getTechnology() {
+    return technology;
+  }
+
+  public void setTechnology(String technology) {
+    this.technology = technology;
+  }
+
+  public String getCompany() {
+    return company;
+  }
+
+  public void setCompany(String company) {
+    this.company = company;
+  }
+
+  public String getDescription() {
+    return description;
+  }
+
+  public void setDescription(String description) {
+    this.description = description;
+  }
+}