+++ src/main/java/org/jallen/tyrael/repositories/WorkRepository.java
@@ -1,0 +1,8 @@
+package org.jallen.tyrael.repositories;
+
+import org.jallen.tyrael.entity.Work;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface WorkRepository extends JpaRepository<Work, Long> {}